Update README.md
# CasinoGuessMaster
We are excited to introduce CasinoGuessMaster, an engaging and dynamic casino-style number guessing game written in C++. As we aim to enhance and expand its features, we are seeking skilled C++ developers to join us in this collaborative effort. Your expertise can help us implement new functionalities, improve existing code, and optimize performance to deliver an even more compelling user experience. By contributing to CasinoGuessMaster, you'll have the opportunity to work on a unique project, share your knowledge with a passionate team, and make a significant impact on the game's development. If you have a passion for C++ programming and are eager to contribute to an innovative gaming project, we would love to hear from you!